+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [0.3.0]
+## [0.3.1]
 
 ### Fixed
 
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 - Rework `CometdError`.
 - `AccessToken` now always set only one token to `authorization` header.
+- `timeout` and `interval` sets through `core::time::Duration`.
 
 ## [0.2.0]